11-serverless-iot-platform/iot-function-app/requirements.txt
azure-functions==1.21.3
azure-cosmos==4.7.0